Rename files for consistent naming
The convention is:
* Classes names are CamelCase.js
* Test files are snake_case_test.js
