+++ test/http_wrapper_test.js
